requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==7.1.2
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
ffmpeg-python==0.2.0
flatbuffers==23.1.4
fonttools==4.38.0
future==0.18.2
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.12.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.8.1
llvmlite==0.38.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
norbert==0.2.1
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pooch==1.6.0
praat-parselmouth==0.4.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
soundfile==0.11.0
spleeter==2.3.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
typer==0.3.2
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
